Generate emails with the correct structure
Now the emails actually have sender and recipients in the correct header
fields, not as key/value pairs in the Content-type header field.

This removes the possibility of having attachments, but that will be
added again later in a different way that lets us specify the attachment
bodies as strings, not files.