Introduce a simple substition mechanism for the mail templates
The read_template function now returns the mail body as an instance of
string.Template. The callers have to call its substitute method with
a suitable mapping to create the actual mail body.

This commit only introduces the mechanism and updates the callers to use
an empty substitution in preparation for format specific substitutions
later.