Fix column order for CSV formats with custom headers
The columns list for the CSV formats may now contain not just strings
but also pairs of strings. The pairs give the intelmq column name and
the column header title for the CSV respectively. This is better than
using a dictionary because the order of the columns is defined by the
format and must not depend on python implementation details.

In addition to the obvious changes necessary for this change in the
column specification, this commit changes the meaning of the header
parameter of format_as_csv to be required and to be a dictionary mapping
dictionary keys to column titles because the one caller currently always
provides this dictionary, so supporting booleans or omitting the
parameter is not needed.

Also, in mail_format_as_csv the columns parameter is not optionaly
anymore and may not be None because all the callers always provide this
value and it's not None.