Tests: Skip a test by default
This introduces a condition to the unit tests which allows certain tests
to be skipped.

This is useful for tests which might only succeed on specific hardware
and fail on development or build machines.

The 'check` Makefile target has been split into 'check' and 'check_all'.
The former is run automatically by build processes, the latter may be
run manually.