load_events does not need to return the list of columns
The list of columns that load_events returns together with the actual
event data is not needed. The callers of load_events either already know
exactly what the columns are if they called load_events with an explicit
list of columns or they will have to assume that they will get all
intelmq attributes. In any case the callers should simply access the
attributes by their names.