Use custom classes to represent the format specifications
The formats are going to become a bit more complicated, particularly in
how the event data has to be transformed into the data in the mails
(e.g. we are going to have to extract information from the extra field).
A simple list of field names and titles will not be enough for that. The
new classes will help with this.

The actual specification still looks very similar to before -- it's
still basically strings, lists and tuples -- because the common should
stay readable.