Debconf: Don't prompt for password twice
Since the password will be written to the config file without any
protection, there is no point in protecting in during the configuration
process.

This has a very useful side effect because, as a simple string, the
password will be cached by debconf.  As the reconfiguration process
rewrites the config file, users would have to reenter the password every
time.  This way, the previous value can simply be confirmed.