Debconf: Improve GPG key import detection
The exit status of `gpg --import` is 2 if the key was already present.
For this package's purpose this is not a failure.

To make sure that the desired key has been imported, no matter when, we
now query gpg with the fingerprint.

Further, we make sure that we're only comparing the fingerprint, not any
other field.  This may very well be overkill here, and can probably
safely be removed.