mailgen: closes #27 and adds asn-grouping.
Supporting that ${asn} can be expanded in the subject line of template files.

One email will be send now per asn within the events. It is also
an example for a decision about the number of emails after loading
the event details in the specific formatting function.

Extend load_events() to include the affected notification_ids, so we
can return them after splitting. This is probably a suboptimal solution.
A better solution would probably remove the early aggregation, keeping the
information there and would thus avoid a second subselect in the
notifactions table.

Adding some debug output to analyse asn-grouping and mark_as_sent().

Refactored some variable names to show where we have aggregated notifications.