Refactored: Move to Python Package structure.
Resolves #19 (see there for the detailed rationale).

Moved to a python package which is a common solution to
be able to split up the functions in several python module files,
test and ship them.

The names have been made more consistent with other intelmq tools,
leaving room for becoming a more general mailer add-on module in the
future.

Technically:

* Git: Ignoring new directories from the python package and pip.
* added a first testcase (without further dependencies)
* added setup.py (and debian/ not updated). Not used intelmq as
  a requirement, because intelmqmail may be used from a different machine.
* Moved requirements to setup.py, avoid another place that needs to be
  maintained. Left out pyxarf requirement for now.
* README: moved the history to the bottom.