Treat a subset of the malware feeds as a single feed
The shadowserver feeds "Botnet-Drone-Hadoop", "Sinkhole-HTTP-Drone" and
"Microsoft-Sinkhole" all contain information about malware and are to be
treated as a single feed for the purposes of intelmqmail. We achieve
this in an ad-hoc way by mapping their feed names to a new synthetic
feed name, "generic_malware", when retrieving the notification
information from the event database and replace the individual CSV
column specifications in feed_specific_formats with a single new one for
"generic_malware". Their CSV formats are exactly the same, so that last
part of the change poses no problems.