Consider classification_type and feed_name for last sent time, too
To honor the notification_interval we need to determine when a
notification of the same type has been sent last. The interesting
question here is what "the same type" means. Since we group notification
by email, template, format, classification_type and feed_name, we should
also consider two notifications to have the same type if they have equal
values for those attributes. The sub-query that determines the newest
sent_at value for a given notification type did not take the last two of
those attributes into account, so this commit adds them.