Remove superfluous print
Printing an error message and immediately before raising an exception
with the same message is usually not necessary or useful. Either the
exception will be shown to the user as well (e.g. by the default
behavior of printing the stacktrace) or the caller handles the
exception, in which case the error message will confuse the user.