Use logging infrastrucuture for all non-interactive output
The one remaining place where print was used when running
non-interactively (i.e. when the --all flag was used) now uses log.info
instead. That message can also be shown in interactive mode but showing
it via log.info instead of print in that case should not be much of a
problem I think, if any.