Do not exit just because some mails could not be generated
When the actual mail creation fails, e.g. because of a missing template,
mailgen should at least try to generate and send the other mails.

With this change any exception raised during mail creation for one
notification is handled by simply logging it and then immediately
proceeding to the next notification.

It does not try to handle exceptions occurring during the actual sending
of the mails or marking of the notifications as sent. If errors happen
during that part, it seems likely that they are of a more permanent
nature, e.g. problems with network connections or more general (i.e. not
notification specific) problems with the SMTP server or the database.