Fetch the event-id/notification-id mapping with the initial query
The ticket numbers are assigned on a per-email basis and which events
are sent in a given email depends on the event attributes. In the
database the ticket numbers are stored in the notifications table,
though, so mailgen must know which notifications belong to which event.

With this commit, the information necessary to determine which
notification belongs to which event is fetched with the initial query
that determines which notifications need to be sent in the first place.
The extra subquery in load_events is not needed anymore. That subquery
also produced incorrect results in cases where one event had multiple
associated notifications because all those notifications were returned,
not just the one that led to that event being included in the email in
question.