Check for the presence of source.asn in the load_events result once
It's not necessary to check for the presence of the source.asn key in
all events returned by load_events. It suffices to check that it's in
the list of columns passed to load_events.

This commit also fixes a defect in the format string for the error
message (%s doesn't work with the format method).