Adapt to the changes in IntelMQ's dev-rework-notification branch
This is first step in the intelmq-mailgen part of the reworking of how
the notification handling works. In both IntelMQ and mailgen this is
handled in the dev-rework-notification branch.

For mailgen this means that it's job is now to simply interpret the
notification directives that IntelMQ added to the events. All the logic
currently present for matching feed.name to email formats is removed.
Instead there's now a rudimentary way to define the formats that the
directives refer to. The example script
example-scripts/10shadowservercsv.py contains the shadowserver formats
previousl directly defined in intelmqmail/cb.py.

The changes affect the database where the notification table has to be
changed. It's now effectively split into the directives table and the
sent table.

This commit introduces some duplication of code present in the
dev-rework-notification branch of IntelMQ, namely the code to load
scripts. We should try to actually share this code later on.