STUDY: A very basic filebased authentication backend.
The backend is capable of emitting JSON Web Tokens (JWT).
These tokens encode a username and the roles associated to
that username.

The study contains two wrapper-functions for pyjwt.
The first consumes a user-object and creates a token,
the second consumes a token and creates a user-object from the
backend.