Define a class for the script entry points
Instead of simply using the function objects directly, we now use class
instances. They are callable, so can be used in practically the same way
as before, but we can more easily add additional information and
functionality. In particular, we now make the filename of the script
available so that we can refer to it in log messages.