Allow scripts to make more decisions
Instead of having the script entry point return a mapping from format
names to format descriptions, the script entry point now gets the full
directive and returns the format description. This allows the script to
actually make decisions about which format to use based on all
information in the directive.

This is probably not flexible enough yet because it's pretty much still
limited to CSV based formats but it's a step in the right direction.

As part of this change in meaning of the entry points the name of the
entry point functions has changed. The example script is updated
accordingly.