Don't repeat column titles in shadowserver example script
Instead of repeating the column titles in every CSV format defined in
the shadowserver example script, we now use a mapping from attribute
name to title and a helper function to assign those titles when creating
all the table format objects.

The main goal of this commit is to show how a script could avoid having
to specify column titles multiple times.