Rework script interfaces to make them more flexible
Instead of having the scripts entry points return a CSV format
description they now return a list of notification objects that are
ready for sending. The actual sending is not done by the scripts so that
steps like the recording of which notifications have been sent is
handled in a uniform way.

The implementation strategy is to provide most of the information a
script needs via a context object that is passed to the script's entry
point function and provides access to the notification directive the
script might want to handle and configuration settings as well as
convenience methods for creating email notifications. The context
implementation contains much of the mail creation code from
intelmqmail/cb.py which has now been moved to the new module
intelmqmail/notification.py which contains the context and notification
implementation.

Due to the changes in the script interface, the entry point has been
renamed to create_notifications. The example script has been updated
accordingly.