Add support for the notification_format directive parameter
This requires changes to the directives table of the event database as
well as the example scripts and some other parts of the code.

As a collateral improvement, this fixes the problem with templates in
xarf directives. The template name was misused for conveying which
variant of xarf to use.