Make the database interaction more robust in case of errors
We have to be careful when errors occur when directives are processed.
In particular, if the errors come from the database and the transaction
is left in a error state that only allows rollbacks we have to make sure
we can still commit the changes made for directives that have been
processed successfully. See the new comment for some more detail.