mail_format_as_csv can now include the CSV data as attachment
The ScriptContext method mail_format_as_csv now has an optional
parameter attach_event_data to indicate whether the CSV formatted event
data should be added to the mail as attachment. The default is false, so
that the default behavior does not change.

The CSV data is available for substitutions in the same way as before
even when attach_event_data is true. Templates used for mails with
attached CSV data should probably not use it, though.

Since signing is now done with OpenPGP/MIME the attachment is part of
the signed data.