Adapt directives_grouping_idx to the actual grouping columns
The columns medium and endpoint are not used when grouping directives
for processing so we remove them from the index. Now it contains exactly
the columns which are used in the GROUP BY clause of the
PENDING_DIRECTIVES_QUERY in intelmqmail/db.py.

The columns medium and endpoint are used elsewhere in the query but
since all rows currently have exactly the values the query tests for, we
don't really profit from an index over those fields anyway.

Now PostgreSQL actually uses the index which makes mailgen substantially
faster!