Allow scripts to decide when the notifications have to be sent
Previously the database query that fetched the pending notifications
from the database also decided for which of them the time for sending
had come (according to the last time a similar notification had been
sent, the current time and the notification interval). Now we simply
make the last sent time available to the scripts so that they can
decide. This gives them more flexibility. For instance, they could
decide to send the notifications earlier than normal if a lot of
notifications have accumulated.

To make it easy to implement the previous behavior, the script context
has a new method that determines whether the notification interval has
been exceeded. Since scripts may now decide to postpone acting on a
directive and postponing is not the same as a script not being able to
act on a directive for some other reason, scripts now also have a new
possible return value, Postponed, that indicates this. Return Postponed
from a script will abort processing of scripts for a directive just like
actually return mails does.

Also, all example scripts have been updated to uses these new facilities
so that they behave just like before.