Fix logic in notification_interval_exceeded
For one, we should use the instance variable that we maintain for the
current time for reasons documented in the comment and, of course, we
should actually do what the doc-string says about a missing sent_at
timestamp meaning that the notification interval has been exceeded.

Add some tests for this. Obviously we should have had some for this in
the first place.