Introduce Directive class for better access to directive parameters
Using a plain dictionary is cumbersome and accessing the directive's
parameters as instance variables makes the interfaces more uniform (less
need to remember which parts need dictionary like syntax and which
don't).

The Dictionary class provides a dictionary like interface for
compatibility, so not all code needs to be adapted immediately. Howerver
this commit adapts the rest of the code in intelmqmail/notification.py
as well because there's a slight incompatibility after all: the
aggregate_identifier attribute is now a dictionary instead of a list of
pairs. Other places in the code that use aggregate_identifier are not
affected.