Simplify initialization of the current time in the script context
The current time for comparisons with notification_interval is also
needed for comparisons with inserted_at and we need to use the same
value for all such comparisons performed for one directive. So instead
of initilizing it on demand in the one method where it's currently used,
we now do it in the constructor and in a simpler way, although the new
way assumes that all timestamps we use are time-zone aware.

Since there's nothing special about its initialization anymore, we might
as well remove the '_'-prefix at the same time.