Add convenience method for dealing with time.observation
When aggregating by time.observation the values stored in the database
are strings. For comparisons with e.g. the current time, they need to be
converted into datetime objects. Also, so far we always want to compare
them with the current time.

So, there's now a function that can convert the values to datetime
objects and a convenience method for ScriptContext that lets scripts
compare the time.observation value with the current time, so that we
don't have to duplicate too much of the logic in the scripts that use
this.