Fix the time.observation comparison in the example scripts
Several example scripts contain logic that compare time.observation and
the current time. With this commit they use the new age_of_observation
method which simplifies the code in the scripts substantially and
reduces code duplication. It also makes the code work, actually, because
now the time.observation value is properly converted to a datetime
object for the comparison.

This also fixes a logic bug in the if statement: For the directive to be
postponed it suffices that one of time.observation or the age of the
newest directive is too small, so we need to use or instead of and.

Also note that if the directives are not actually aggregated by
time.observation the new implementation will raise an exception. This
should be OK since it should be considered a defect in either the
scripts of the rule expert bot or the mailgen scripts if this happens.