Improve update instructions for the new inserted_at field
The instructions now take into account that new directives may be
inserted while the directives table is upgraded. This is done by setting
the default value of the inserted_at column before filling in the
inserted_at column of existing tuples with the event's time.observation
value. After this step no tuple should have a NULL value in inserted_at
any more so we can add the NOT NULL constraint.