Unsigned mails should only have one level of multipart/mixed
In the case of unsigned mails with attachments, the mail itself should
have content-type multipart/mixed cotaining a text/plain part and one
additional part for each attachment. Due to a defect, the
multipart/mixed part was contained in an additional multipart/mixed
wrapper.

The case of OpenPGP/MIME signed mails was handle correctly, though.