Perform "From "-escaping even if the quoted-printable was requested
For some reason the code in FromQuotingContentManager.set_content does
not attempt to quote "From " at the beginning of a line if the caller
has explicitly requested quoted-printable as a
content-transfer-encoding. This is wrong, because clearly the text could
contain a problematic "From " even if the caller requested
quoted-printable. So this even checks for the need to quote the "From "
if the caller requested quoted-printable.