Enforce quoted-printable for text MIME parts
The reason for this is that this makes sure that text is sent as pure
ASCII and with reasonable short lines. We do this in mailgen's custom
ContentManager which now serves two purposes: enforcing quoted-printable
and quoting "From " (as before). The latter is now simpler because we
can assume that all text is quoted printable and we can simply always do
the quoting. We still need to make sure that we only do this for text,
but checking that the actual transfer encoding is quoted-printable
suffices for that.

The content manager is now actually about more than escaping "From ", so
we rename it to the more general MailgenContentManager. The custom email
policy is now also about more than just making sure OpenPGP/MIME
signatures work correctly, so rename it as well to mailgen_policy.