Make sending notification more robust when SMTP fails
This commit swaps the nesting of the loop over the directives and the
context manager for the SMTP connection so that the directive loop is
outside of the SMTP context manager. This makes mailgen more robust
because it can now recover from a problem with the SMTP connection.

If the SMTP connection fails, a new connection will be opened for the
next (group of) directives which, if it succeeds, will allow the
corresponding notifications to be sent. With the SMTP context manager
outside of the directive loop, a failure in open SMTP connection would
lead to mailgen not being able to process any of the subsequent
directives at all.