Explicitly generate a Message-Id for notification mails.
The MTA we're handing the mails off to does not necessarily add a
Message-Id, so we now do that explicitly. The domain part of the
Message-Id is taken from the sender address because we cannot rely on
the default domain name chosen by the library function we use to
generate the Message-Id (email.utils.make_msgid uses the system's
hostname but that may not be configured correctly).