Make it easier to use items from the extra field in table formats
So far, scripts that wanted to include information from the extra field
the table with the event data had to directly use the ExtraColumn class.
With this commit they can just pretend that the keys of the dictionary
in the extra field are normal event fields that have a "extra." prefix.

E.g. when using build_table_formats to create the table format
definitions, instead of having to specify a column with ExtraColumn like
this:

    ExtraColumn("mssql_version", "version")

one can now use this:

    ("extra.version", "mssql_version")

The ordering of the parameters, with ExtraColumn it's title followed by
key, with the tuple it's the other way around, is due to historical
reasons and not addressed by this commit.

The change for this commit is in build_table_column, so it indirectly
affects the higher level functions build_table_format and
build_table_formats as well.