Allow passing a logger instance to load_scripts
When used from within a bot in IntelMQ, the default logger used by
load_scripts does not actually log anything. This appears to be due to
the way the logger is instantiated in IntelMQ, but I haven't checked
that completely. When passing the bot's logger explicitly to
load_scripts script, logging works as expected.

This commit also modifies one of the load_scripts test cases to pass a
logger explicitly. Another still does not provide one, so that both
cases, an explicit logger and the implicit logger are tested.