Allow log-level to be set for the intelmqmail.cb logger, too
Users aresupposed to be able to set the log-level logging_level option
in the configuration file. This is implemented by setting that level on
the logger "intelmqmail" which is the parent for all mailgen specific
logger and therefore should influence all of them. However, the module
intelmqmail.cb set the level of its own logger explicitly to INFO, which
meant that this had precence over the log-level of the parent, and
therefore the log-level set in the configuration file did not apply to
intelmqmail.cb.

With this commit, the cb module does not explicitly assign a level to
its logger anymore and the code in main where the log-level for all of
mailgen is set always sets the loglevel -- not just in the case where
there's a log-level in the config file -- and uses INFO as the default
level if no level is specified in the configuration.