To improve speed, determine last_sent over the last month only
The main query used by mailgen determines the notification directives
that have not yet been processed and some related data. Part of this is
a subquery that determines when similar notifications have last been
sent so that the mailgen scripts can decide whether enough time has
passed since then for new notifications to be sent.

This sub-query turns out to be quite slow in some cases, so that
simetimes the whole query takes much too long (up to several hours).
Limiting the subquery to only consider similar directives from the last
4 weeks improves this query substantially. The main factor that
influences the speed is the number of similar directives already in the
table.

It's not clear yet whether this commit improves speed enough for
production installations, but it should improve the situation.

The value of 4 weeks was chosen based on practical needs but is still
somewhat arbitrary. The main criterion for this value is that it must be
greater than the maximum expected notification interval because the
last_sent date will never be older than 4 weeks now.