Add inserted_at to the grouping index
The query mailgen uses to determine the directives to be processed, now
also uses the inserted_at column. That alone speeds up the query
substantially, but it's even better if that column is also in the index
PostgreSQL uses in that part of the query. That index is
directives_grouping_idx. My Tests indicate that without the inserted_at
column in the index, the speedup is about 25%, with the column in the
index its >80%.

To avoid confusion, the index with the added column is now called
directives_grouping_inserted_at_idx. It also makes the upgrade perhaps a
little easier because we can just add the new index and then remove the
old one.