Show signature rate when the signing speed test fails
The test suite has a test that determines the rate at which OpenPGP
signatures can be performed. That test is not run by default, but when
it is run is useful to know the rate when the test fails because it's
too low. By using assertGreater instead of assertTrue we get that
automatically from Python's unittest module.