Store timestamp of email Date header in sent
The sent_at column of the sent table records when the notifications were
sent. Until now the value of sent_at ended up being same for all mails
sent during one mailgen run because we used PostgreSQL now() function
which gives the same value for all calls during a single transaction.

Now we use the timestamp that was used to create the Date header of the
emails that are sent. This is different for each mail and more
accurately reflects the actual time when the mail was sent.

I've tested that this handles time zones correctly by running mailgen
with an explicit timezone (e.g. TZ=Europe/Berlin mailgen ...) in which
case the Date header of the mails contains the correct timezone
information (e.g. "Date: Tue, 09 Jul 2019 15:34:21 +0200") and the
database also has the right time stamp (e.g. in this case in UTC:
2019-07-09 13:34:21+00).