Assume newest inserted_at correlates to newest sent_at
To improve the speed of the main query that determines the directives
that need processing, we now rely on a correlation between
directives.inserted_at and sent.sent_at when determining when the last
similar mail was sent. We assume that the newest directive in the set of
all directives with the same aggregation identifiert that have already
been processed belongs to the newest such email. With the way mailgen is
used and uses the database and the way events/directives are inserted
into the database, this assumption should hold.