Allow display names in sender.
mailgen determines the domain for the Message-ID in the mails it send
from the sender address that was configured in the config file. So far,
we assumed that that was given in the "user@domain" form when extracting
the domain from it. This makes it impossible to configure a display name
for the From header by simply specifying sender in the form "displayname
<user@domain>".

We now parse the sender properly as if it were a From: header value with
the new helper function domain_from_sender and use this helper in the
two places were the domain was extract from sender.

There are also two new test cases that check whether the Message-IDs
generated from the sender are syntactically correct. They weren't with
the old code when the sender contained a display name and the email
surrounded by angle brackets.