Give python3-gpg as setup.py dependency
Using the environment markers of the requirement specification
to make a distinction between elder python versions were python3-gpgme
was availabe in Debian and Ubuntu and newer version where we want to
support python3-gpg. This is a pragmatic approach as pip 8.1.1
(coming with Ubuntu 16.04 LTS) does not fully recognise the environment
marker as we use them.

improve #44